+++ lib/src/core/escape/parser.dart
@@ -206,7 +206,7 @@ class EscapeParser {
 
   /// The last parsed [_Csi]. This is a mutable singletion by design to reduce
   /// object allocations.
-  final _csi = _Csi(finalByte: 0, params: []);
+  final _csi = _Csi(finalByte: 0, params: [], intermediates: []);
 
   /// Parse a CSI from the head of the queue. Return false if the CSI isn't
   /// complete. After a CSI is successfully parsed, [_csi] is updated.
@@ -216,6 +216,7 @@ class EscapeParser {
     }
 
     _csi.params.clear();
+    _csi.intermediates.clear();
 
     // test whether the csi is a `CSI ? Ps ...` or `CSI Ps ...`
     final prefix = _queue.peek();
@@ -252,7 +253,7 @@ class EscapeParser {
       }
 
       if (char > Ascii.NULL && char < Ascii.num0) {
-        // intermediates.add(char);
+        _csi.intermediates.add(char);
         continue;
       }
 
@@ -278,6 +279,7 @@ class EscapeParser {
     'l'.codeUnitAt(0): _csiHandleMode,
     'm'.codeUnitAt(0): _csiHandleSgr,
     'n'.codeUnitAt(0): _csiHandleDeviceStatusReport,
+    'q'.codeUnitAt(0): _csiHandleCursorStyle,
     'r'.codeUnitAt(0): _csiHandleSetMargins,
     't'.codeUnitAt(0): _csiWindowManipulation,
     'A'.codeUnitAt(0): _csiHandleCursorUp,
@@ -298,6 +300,19 @@ class EscapeParser {
     'X'.codeUnitAt(0): _csiHandleEraseCharacters,
     '@'.codeUnitAt(0): _csiHandleInsertBlankCharacters,
   });
+
+  void _csiHandleCursorStyle() {
+    if (_csi.intermediates.length != 1 ||
+        _csi.intermediates.single != Ascii.space) {
+      return handler.unknownCSI(_csi.finalByte);
+    }
+
+    final style = switch (_csi.params) {
+      [final style] => style,
+      _ => 0,
+    };
+    handler.setCursorShape(style);
+  }
 
   /// `ESC [ Ps a` Cursor Horizontal Position Relative (HPR)
   ///
@@ -1131,7 +1146,7 @@ class _Csi {
   _Csi({
     required this.params,
     required this.finalByte,
-    // required this.intermediates,
+    required this.intermediates,
   });
 
   int? prefix;
@@ -1139,7 +1154,8 @@ class _Csi {
   List<int> params;
 
   int finalByte;
-  // final List<int> intermediates;
+
+  final List<int> intermediates;
 
   @override
   String toString() {