+++ lib/src/ui/render.dart
@@ -431,7 +431,7 @@ class RenderTerminal extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
         _painter.paintCursor(
           canvas,
           offset + cursorOffset,
-          cursorType: _cursorType,
+          cursorType: _terminal.applicationCursorType ?? _cursorType,
           hasFocus: _focusNode.hasFocus,
         );
       }