+++ test/src/core/escape/parser_test.mocks.dart
@@ -421,6 +421,14 @@ class MockEscapeHandler extends _i1.Mock implements _i2.EscapeHandler {
         returnValueForMissingStub: null,
       );
   @override
+  void setCursorShape(int? style) => super.noSuchMethod(
+        Invocation.method(
+          #setCursorShape,
+          [style],
+        ),
+        returnValueForMissingStub: null,
+      );
+  @override
   void setInsertMode(bool? enabled) => super.noSuchMethod(
         Invocation.method(
           #setInsertMode,